test(client): improve Codecov coverage and test descriptions
