+++ codecov.yml
@@ -1,0 +1,50 @@
+codecov:
+  notify:
+    after_n_builds: 5
+
+coverage:
+  status:
+    project:
+      default:
+        target: auto
+        threshold: 1%
+    patch:
+      default:
+        target: auto
+        threshold: 1%
+
+comment:
+  layout: "header, diff, flags, components"
+  require_changes: true
+
+component_management:
+  default_rules:
+    statuses:
+      - type: project
+        target: auto
+        threshold: 1%
+  individual_components:
+    - component_id: cb_events_client
+      name: client
+      paths:
+        - src/cb_events/client.py
+    - component_id: cb_events_models
+      name: models
+      paths:
+        - src/cb_events/models.py
+    - component_id: cb_events_router
+      name: router
+      paths:
+        - src/cb_events/router.py
+    - component_id: cb_events_config
+      name: config
+      paths:
+        - src/cb_events/config.py
+    - component_id: cb_events_exceptions
+      name: exceptions
+      paths:
+        - src/cb_events/exceptions.py
+    - component_id: cb_events_public_api
+      name: package_api
+      paths:
+        - src/cb_events/__init__.py
