+++ pyproject.toml
@@ -203,13 +203,12 @@ addopts = [
     "--strict-markers",
     "--strict-config",
     "--tb=short",
-    "--cov-report=term-missing:skip-covered",
-    "--cov=src",
     "-ra",
 ]
 markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
     "e2e: marks tests requiring external services",
+    "live: marks tests requiring explicit opt-in live credentials",
 ]
 
 [tool.coverage.run]
