+++ tests/client/test_request_retry.py
@@ -66,7 +66,7 @@ async def test_testing_mode_caps_attempts(
     testbed_url_pattern: re.Pattern[str],
     is_exception: bool,
 ) -> None:
-    """When testing mode is enabled, the client should stop retrying after the specified number of attempts, even if retry conditions persist."""
+    """Testing mode should cap retries even when failures keep occurring."""
     if is_exception:
         aioresponses_mock.get(
             testbed_url_pattern,
@@ -99,7 +99,7 @@ async def test_retries_on_retry_status_codes_then_succeeds(
     aioresponses_mock: aioresponses,
     testbed_url_pattern: re.Pattern[str],
 ) -> None:
-    """Client should retry on HTTP status codes in RETRY_STATUS_CODES and eventually return events when a subsequent request succeeds."""
+    """Client should retry retryable statuses and eventually succeed."""
     success_response = make_response([make_event(EventType.TIP, event_id="1")])
     aioresponses_mock.get(testbed_url_pattern, status=502)
     aioresponses_mock.get(testbed_url_pattern, payload=success_response)
